fix: upload profile on first sign-in
This fixes 2 issues, one is that the Photo Selector dialog was open after selecting or taking a photo. This has been resolved by adding RouteSettings and using .popUntil. The second issue was that it was not using the correct endpoint, seems it was some leftover from when we were using firebase.